pyproject.toml
[project]
name = "reverb_analyzer"
version = "0.1.0"
description = "A podcast analysis and newsletter generation system"
authors = [
    {name = "kavwad",email = "123405488+kavwad@users.noreply.github.com"},
    {name = "patrickswnsn",email = "patrickdswanson@icloud.com"},
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts (>=0.2.1,<0.3.0)",
    "google-generativeai (>=0.8.3,<0.9.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "requests (>=2.32.3,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "lxml (>=5.3.0,<6.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "pytz (>=2024.2,<2025.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "src" },
    { include = "core", from = "src" },
    { include = "utils", from = "src" },
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
ipykernel = "^6.29.5"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.10"
pytest-asyncio = "^0.25.3"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
select = ["E", "F", "B", "I"]
ignore = []
line-length = 88
target-version = "py313"

